moved some functions out into separate .py files. (#3937)
